pivoted to just piping the straight analog signal, interpolation to exact distance was not needed, saved a few ops

+++ cc1310_code/dist_sense.h
@@ -11,7 +11,7 @@
 #include "adc.h"
 #include <stdint.h>
 
-float ReadDist();
+uint32_t ReadDist();
 
 
 #endif /* DIST_SENSE_H_ */
